fix: 清除 clang-tidy 全量 0 警告，types.h/error_code.h 移入 src/
- include/ 目录移除，types.h、error_code.h 归入 src/
- types.h: typedef→using, _Types→Types, [[nodiscard]], range-for,
  braced-init, #include <xstring>→<string>, NOLINT 宏误报, frozen 注释
- node.h: getFullName [[nodiscard]], Named pass-by-value
- ast_builder.h: currentScope/findEnclosingStructName [[nodiscard]]
- statement_node.h: 移除 ptr std::move, Token move, AssignOp:u8
- compiler_destructor.cpp: 合并 isDynBorrow/isHeap 空分支
- compiler.cpp: 指定初始化 TestFnInfo
- compiler_expr.cpp: NOLINT createCast/compileExpr branch-clone 误报
- expr_ctor.cpp: const auto& 避免拷贝
- syntax_error_listener.cpp: int→size_t 符号比较
- symbol_lookup.cpp: fopen→fopen_s, NOLINT empty-catch
- docs/dev/*, docs/spec/draft/*: include/types.h→src/types.h
- .claude/rules/directory.md: 移除 include/ 条目

+++ src/tools/syntax_error_listener.cpp
@@ -3,8 +3,8 @@
 
 #include "syntax_error_listener.h"
 
+#include "../error_code.h"
 #include "diagnostic.h"
-#include "error_code.h"
 
 #include "Lexer.h"
 #include "TokenStream.h"
@@ -96,7 +96,7 @@ void SyntaxErrorListener::syntaxError(antlr4::Recognizer* recognizer,
             auto* ts = dynamic_cast<antlr4::TokenStream*>(recognizer->getInputStream());
             if (ts && offending) {
                 auto idx = offending->getTokenIndex();
-                for (int i = 1; i <= 5 && idx >= i; ++i) {
+                for (size_t i = 1; i <= 5 && idx >= i; ++i) {
                     auto* tok = ts->get(idx - i);
                     if (tok && tok->getChannel() == antlr4::Token::DEFAULT_CHANNEL) {
                         prevText = tok->getText();
